progess
some stuff is working some stuff is not everything is horrible...

+++ Tier3/src/main/java/via/sep3/group11/tier3/protobuf/GHousingOrBuilder.java
@@ -29,8 +29,8 @@ public interface GHousingOrBuilder extends
   via.sep3.group11.tier3.protobuf.GAddressOrBuilder getAddressOrBuilder();
 
   /**
-   * <code>int32 id = 3;</code>
+   * <code>int64 id = 3;</code>
    * @return The id.
    */
-  int getId();
+  long getId();
 }
